+++ source/LibUISharp/src/LibUISharp/Drawing/FillMode.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Specifies how the interior of a closed path is filled.
     /// </summary>
-    public enum FillMode : uint
+    public enum FillMode : long
     {
         /// <summary>
         /// Specifies the winding fill mode.
