+++ source/LibUISharp/src/LibUISharp/Drawing/KeyModifierFlags.cs
@@ -4,7 +4,7 @@ namespace LibUISharp.Drawing
 {
     // uiModifier
     [Flags]
-    public enum KeyModifierFlags : uint
+    public enum KeyModifierFlags : long
     {
         Ctrl = 1 << 0,
         Alt = 1 << 1,
