+++ source/LibUISharp/src/LibUISharp/Drawing/WindowEdge.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Specifies the edge of a <see cref="Window"/> object.
     /// </summary>
-    public enum WindowEdge : uint
+    public enum WindowEdge : long
     {
         /// <summary>
         /// Specifies the left edge of a <see cref="Window"/>.
